@@ -24,19 +24,25 @@ class SleepsController < ApplicationController
               }
             ]
           }
-      EOS
+    EOS
     api_base_url ""
     formats ['html', 'json']
   end
 
   def_param_group :sleeps_data do
-    param :start_time, :undef, :desc => "Sleep Start Time [DateTime(UTC)]", :required => true
-    param :finish_time, :undef, :desc => "Sleep Finish Time", :required => true
-    param :quality, Integer, :desc => "[['Broken', 1], ['Light', 2], ['Normal', 3], ['Deep',4]]", :required => true
+    param :sleep, Hash, :desc => "Sleep", :required => false do
+      param :start_time, :undef, :desc => "Sleep Start Time [DateTime(UTC)]", :required => false
+      param :finish_time, :undef, :desc => "Sleep Finish Time [DateTime(UTC)]", :required => false
+      param :quality, :number, :desc => "[['Broken', 1], ['Light', 2], ['Normal', 3], ['Deep',4]]", :required => true
+    end
   end
 
   def_param_group :sleeps_all do
     param_group :sleeps_data
+  end
+
+  def_param_group :destroy_sleeps_data do
+    param :id, :number, :desc => "Id of Sleep Entry to Delete [Number]", :required => true
   end
 
   acts_as_token_authentication_handler_for User
@@ -45,14 +51,15 @@ class SleepsController < ApplicationController
   before_action :set_lookback_period, only: [:index]
   before_action :authenticate_user!
 
-  after_filter :verify_authorized,  except: [:index]
+  after_filter :verify_authorized, except: [:index]
   #after_filter :verify_policy_scoped, only: [:index]
 
-  respond_to :js
-  
+  respond_to :html, :js, :json
+
   # GET /sleeps
   # GET /sleeps.json
   api! "Show Sleep Entries"
+
   def index
     @user = User.find_by_id(params[:user_id])
 
@@ -74,7 +81,6 @@ class SleepsController < ApplicationController
       format.html
       format.js
       format.json { render :json => @sleeps, status: 200 }
-      format.xml { render :xml => @sleeps, status: 200 }
     end
   end
 
@@ -85,9 +91,8 @@ class SleepsController < ApplicationController
     authorize! :read, Sleep
 
     respond_to do |format|
-      format.js
-      format.json { render :json =>  @sleep, status: 200 }
-      format.xml { render :xml => @sleep, status: 200 }
+      format.js { render :nothing => true }
+      format.json { render :json => @sleep, status: 200 }
     end
   end
 
@@ -106,11 +111,15 @@ class SleepsController < ApplicationController
     end
 
     @sleep= Sleep.new
+
+    respond_to do |format|
+      format.html { render :nothing => true }
+      format.js
+      format.json { render :json => @sleep, status: 200 }
+    end
   end
 
   # GET /sleeps/1/edit
-  api! "Edit Sleep Entry"
-  param_group :sleeps_all
   def edit
     @user = User.find_by_id(params[:user_id])
 
@@ -123,12 +132,19 @@ class SleepsController < ApplicationController
         authorize @sleeps
       end
     end
+
+    respond_to do |format|
+      format.html { render :nothing => true }
+      format.js
+      format.json { render :json => @sleep, status: 200 }
+    end
   end
 
   # POST /sleeps
   # POST /sleeps.json
   api! "Create Sleep Entry"
   param_group :sleeps_data
+
   def create
     @user = User.find_by_id(params[:user_id])
 
@@ -145,17 +161,27 @@ class SleepsController < ApplicationController
     @sleep = Sleep.new(sleep_params)
     @sleep.user_id = current_user.id
     @sleep.time = (@sleep.finish_time.to_i - @sleep.start_time.to_i) / 3600
-    
+
+    #days_sleeps = Sleep.where(user_id: current_user.id, finish_time: (@sleep.finish_time.to_time.in_time_zone.at_beginning_of_day..@sleep.finish_time.to_time.in_time_zone.end_of_day))
+    days_sleeps = Sleep.where(user_id: current_user.id, finish_time: ((@sleep.finish_time.to_time.in_time_zone.- 24.hours)..@sleep.finish_time.to_time.in_time_zone))
+
     respond_to do |format|
-      if @sleep.save
-        track_sleep_created
-        current_user.scorecard.update_scorecard('sleeps')
-        flash.now[:success] = 'Sleep Entry was successfully tracked.'
-        format.js 
-        format.json { render :json => @sleep, status: :created }
+      if days_sleeps.count < MAX_SLEEP_ENTRIES
+        if @sleep.save
+          track_sleep_created
+          current_user.scorecard.update_scorecard('sleeps', Time.zone.now)
+          flash.now[:success] = 'Sleep Entry was successfully tracked.'
+          format.js
+          format.json { render :json => @sleep, status: :created }
+        else
+          flash.now[:error] = 'Sleep Entry was not tracked... Try again???'
+          format.js { render json: @sleep.errors, status: :unprocessable_entity }
+          format.json { render json: @sleep.errors, status: :unprocessable_entity }
+        end
       else
-        format.js { render json: @sleep.errors, status: :unprocessable_entity }
-        format.json { render json: @sleep.errors, status: :unprocessable_entity }
+        flash.now[:warning] = 'Sleep Entry was not tracked.  Daily Sleep Entry Limit Reached.'
+        format.js
+        format.json { render json: 'Sleep Entry was not tracked.  Daily Sleep Entry Limit Reached.', status: 400 }
       end
     end
   end
@@ -164,6 +190,7 @@ class SleepsController < ApplicationController
   # PATCH/PUT /sleeps/1.json
   api! "Update Sleep Entry"
   param_group :sleeps_all
+
   def update
     @user = User.find_by_id(params[:user_id])
 
@@ -177,6 +204,10 @@ class SleepsController < ApplicationController
       end
     end
 
+    finish_time = sleep_params[:finish_time].to_datetime
+    #days_sleeps = Sleep.where(user_id: current_user.id, finish_time: (finish_time.in_time_zone.at_beginning_of_day..finish_time.in_time_zone.end_of_day))
+    days_sleeps = Sleep.where(user_id: current_user.id, finish_time: ((finish_time.to_time.in_time_zone.- 24.hours)..finish_time.to_time.in_time_zone))
+
     respond_to do |format|
       if @sleep.update(sleep_params)
         @sleep.time = (@sleep.finish_time.to_i - @sleep.start_time.to_i) / 3600
@@ -184,12 +215,12 @@ class SleepsController < ApplicationController
         track_sleep_updated
 
         flash.now[:success] = 'Sleep Entry was successfully updated.'
-        format.js
+        format.js { render status: 200 }
         format.json { render :json => @sleep, status: :created }
       else
-        flash.now[:error] = 'Sleep Entry was not updated... Try again???'
-        format.js { render json: @sleep.errors, status: :unprocessable_entity }
-        format.json { render json: @sleep.errors, status: :unprocessable_entity }
+        flash.now[:error] = 'Sleep Entry was not updated.  Daily Sleep Entry Limit Reached.'
+        format.js
+        format.json { render json: 'Sleep Entry was not updated.  Daily Sleep Entry Limit Reached.', status: 400 }
       end
     end
   end
@@ -213,6 +244,8 @@ class SleepsController < ApplicationController
   # DELETE /sleeps/1
   # DELETE /sleeps/1.json
   api! "Delete Sleep Entry"
+  param_group :destroy_sleeps_data
+
   def destroy
     @user = User.find_by_id(params[:user_id])
 
@@ -225,14 +258,20 @@ class SleepsController < ApplicationController
         authorize @sleeps
       end
     end
-    
-    @sleep.destroy
-    track_sleep_deleted
 
     respond_to do |format|
-      flash.now[:success] = 'Sleep Entry was successfully removed.'
-      format.js
-      format.json { head :no_content }
+      if @sleep.destroy
+        track_sleep_deleted
+        flash[:success] = 'Sleep Entry was successfully deleted.'
+        format.html { redirect_to sleeps_path }
+        format.js
+        format.json { head :no_content }
+      else
+        flash[:error] = 'Sleep Entry was not deleted... Try again???'
+        format.html { redirect sleeps_path }
+        format.js
+        format.json { render json: @sleeps.errors, status: :unprocessable_entity }
+      end
     end
   end
 
@@ -248,66 +287,70 @@ class SleepsController < ApplicationController
         authorize @sleeps
       end
     end
-    
+
+    $current_capture_screen = "Sleep"
+
     respond_to do |format|
       format.js
     end
   end
 
   private
-    # Use callbacks to share common setup or constraints between actions.
-    def set_sleep
-      @sleep = Sleep.find(params[:id])
-    end
+  # Use callbacks to share common setup or constraints between actions.
+  def set_sleep
+    @sleep = Sleep.find(params[:id])
+  end
 
-    def set_lookback_period
-      if(params.has_key?(:sleep_lookback_period))
-        @sleep_lookback_period = params[:sleep_lookback_period]
-      else
-        @sleep_lookback_period = DEFAULT_LOOKBACK_PERIOD
-      end
+  def set_lookback_period
+    if params.has_key? :sleep_lookback_period
+      @sleep_lookback_period = params[:sleep_lookback_period]
+    else
+      @sleep_lookback_period = DEFAULT_LOOKBACK_PERIOD
     end
+  end
 
-    # Never trust parameters from the scary internet, only allow the white list through.
-    def sleep_params
-      params.fetch(:sleep, {}).permit(:start_time, :finish_time, :quality, :sleep_lookback_period)
-    end
+  # Never trust parameters from the scary internet, only allow the white list through.
+  def sleep_params
+    params.fetch(:sleep, {}).permit(:start_time, :finish_time, :quality, :sleep_lookback_period)
+  end
 
-    def track_sleep_created
-      # Track Sleep Creation for Segment.io Analytics
-      Analytics.track(
-        user_id: @sleep.user_id,
-        event: 'Created Sleep Entry',
-        properties: {
-          sleep_id: @sleep.id,
-          start_time: @sleep.start_time,
-          finish_time: @sleep.finish_time,
-          quality: @sleep.quality
-        }
-      )
-    end
+  def track_sleep_created
+    # Track Sleep Creation for Segment.io Analytics
+    Analytics.track(
+      user_id: current_user.id,
+      event: 'Sleep Entry Created',
+      properties: {
+        sleep_id: @sleep.id,
+        start_time: @sleep.start_time,
+        finish_time: @sleep.finish_time,
+        quality: @sleep.quality,
+        sleep_user_id: @sleep.user_id
+      }
+    )
+  end
 
-    def track_sleep_updated
-      # Track Sleep Update for Segment.io Analytics
-      Analytics.track(
-        user_id: @sleep.user_id,
-        event: 'Updated Sleep Entry',
-        properties: {
-          sleep_id: @sleep.id,
-          start_time: @sleep.start_time,
-          finish_time: @sleep.finish_time,
-          quality: @sleep.quality
-        }
-      )
-    end
+  def track_sleep_updated
+    # Track Sleep Update for Segment.io Analytics
+    Analytics.track(
+      user_id: current_user.id,
+      event: 'Sleep Entry Updated',
+      properties: {
+        sleep_id: @sleep.id,
+        start_time: @sleep.start_time,
+        finish_time: @sleep.finish_time,
+        quality: @sleep.quality,
+        sleep_user_id: @sleep.user_id
+      }
+    )
+  end
 
-    def track_sleep_deleted
-      # Track Sleep Deletion for Segment.io Analytics
-      Analytics.track(
-        user_id: @sleep.user_id,
-        event: 'Deleted Sleep Entry',
-        properties: {
-        }
-      )
-    end
+  def track_sleep_deleted
+    # Track Sleep Deletion for Segment.io Analytics
+    Analytics.track(
+      user_id: current_user.id,
+      event: 'Sleep Entry Deleted',
+      properties: {
+      }
+    )
+  end
 end
