@@ -10,18 +10,39 @@
 // Read Sprockets README (https://github.com/sstephenson/sprockets#sprockets-directives) for details
 // about supported directives.
 //
-//= require jsapi
 //= require chartkick
 //= require jquery
-//= require jquery-ui/button
-//= require jquery-ui/datepicker
-//= require jquery-ui/spinner
-//= require jquery-ui/tooltip
-//= require jquery-ui/effect
-//= require jquery-ui/slider
+//= require jquery-ui
 //= require jquery_ujs
+//= require sharrre
+//= require twitter/bootstrap
+// require flatui
+// require startup_framework
 //= require jquery.slick
-//= require bootstrap.min.js
-//= require flatui
-//= require startup_framework
+//= require chosen-jquery
 //= require slick
+//= require messages
+// require bootstrap-material-design
+//= require cocoon
+//= require analytics
+//= require custom_overrides
+
+(function () {
+    var qs, js, q, s, d = document, gi = d.getElementById, ce = d.createElement, gt = d.getElementsByTagName, id = 'typef_orm', b = 'https://s3-eu-west-1.amazonaws.com/share.typeform.com/';
+    if (!gi.call(d, id)) {
+        js = ce.call(d, 'script');
+        js.id = id;
+        js.src = b + 'share.js';
+        q = gt.call(d, 'script')[0];
+        q.parentNode.insertBefore(js, q)
+    }
+    id = id + '_';
+    if (!gi.call(d, id)) {
+        qs = ce.call(d, 'link');
+        qs.rel = 'stylesheet';
+        qs.id = id;
+        qs.href = b + 'share-button.css';
+        s = gt.call(d, 'head')[0];
+        s.appendChild(qs, s)
+    }
+})()
