@@ -1,7 +1,7 @@
-//= require ../startup-framework/flat-ui/js/jquery.ui.touch-punch.min
-//= require ../startup-framework/flat-ui/js/bootstrap-select
-//= require ../startup-framework/flat-ui/js/bootstrap-switch
-//= require ../startup-framework/flat-ui/js/flatui-checkbox
-//= require ../startup-framework/flat-ui/js/flatui-radio
-//= require ../startup-framework/flat-ui/js/jquery.tagsinput
-//= require ../startup-framework/flat-ui/js/jquery.placeholder
+//= require jquery.ui.touch-punch.min
+//= require bootstrap-select
+//= require bootstrap-switch
+//= require flatui-checkbox
+//= require flatui-radio
+//= require jquery.tagsinput
+//= require jquery.placeholder
