@@ -1,3 +1,3 @@
 class SuperusersController < ApplicationController
-	before_action :authenticate_user!
+  before_action :authenticate_user!
 end
