@@ -39,13 +39,13 @@ class Users::PasswordsController < Devise::PasswordsController
 
   protected
 
-    def after_resetting_password_path_for(resource)
-      # super(resource)
-      new_session_path(resource)
-    end
+  def after_resetting_password_path_for(resource)
+    # super(resource)
+    new_session_path(resource)
+  end
 
-    # The path used after sending reset password instructions
-    def after_sending_reset_password_instructions_path_for(resource_name)
-      super(resource_name)
-    end
+  # The path used after sending reset password instructions
+  def after_sending_reset_password_instructions_path_for(resource_name)
+    super(resource_name)
+  end
 end
