@@ -7,9 +7,11 @@ class WelcomeController < ApplicationController
         # redirect_to :controller => 'moods', :action => 'index'
         render "/superusers/index.html", :layout => "application"
       elsif user_signed_in? && (current_user.is? :admin)
-        render "/admins/index.html",:layout => "application"
+        render "/admins/index.html", :layout => "application"
       elsif user_signed_in? && (current_user.is? :pro)
-        render "/pros/index.html", :layout => "application"
+        redirect_to pros_path, :layout => "application"
+      elsif user_signed_in? && (current_user.is? :supporter)
+        redirect_to supporters_path, :layout => "application"
       elsif user_signed_in? && (current_user.is? :client)
         render "/clients/index.html", :layout => "application"
       elsif user_signed_in? && (current_user.is? :user)
