@@ -1,3 +1,5 @@
 class Journal < ActiveRecord::Base
-	belongs_to :user
+  belongs_to :user
+
+  validates_presence_of :journal_entry, :timestamp
 end
