@@ -1,5 +1,7 @@
 class Reward < ActiveRecord::Base
-	belongs_to :user
+  belongs_to :user
+
+  validates_presence_of :user_id, :rewards_enabled
 
   def set_rewards_enabled(bool_value)
     self.rewards_enabled = bool_value
