@@ -13,11 +13,11 @@ Rails.application.configure do
   config.eager_load = false
 
   # Configure static asset server for tests with Cache-Control for performance.
-  config.serve_static_assets  = true
+  config.serve_static_assets = true
   config.static_cache_control = 'public, max-age=3600'
 
   # Show full error reports and disable caching.
-  config.consider_all_requests_local       = true
+  config.consider_all_requests_local = true
   config.action_controller.perform_caching = false
 
   # Raise exceptions instead of rendering exception templates.
@@ -39,5 +39,9 @@ Rails.application.configure do
 
   # Mandrill Settings
   # ActionMailer Config
-  config.action_mailer.default_url_options = { :host => 'thrivestreams.com' }
+  config.action_mailer.default_url_options = {:host => 'thrivestreams.com'}
+
+  # Mailboxer Dependency Issue
+  # https://github.com/mailboxer/mailboxer/issues/268
+  config.active_record.maintain_test_schema = false
 end
