@@ -33,7 +33,7 @@ SimpleTokenAuthentication.configure do |config|
   #   Then both the header names identifier key and default value are modified accordingly:
   #     `config.header_names = { super_admin: { phone_number: 'X-SuperAdmin-PhoneNumber' } }`
   #
-  config.header_names = { user: { authentication_token: 'X-User-Token', email: 'X-User-Email' } }
+  config.header_names = {user: {authentication_token: 'X-User-Token', email: 'X-User-Email'}}
 
   # Configure the name of the attribute used to identify the user for authentication.
   # That attribute must exist in your model.
@@ -51,7 +51,7 @@ SimpleTokenAuthentication.configure do |config|
   #
   #   `config.identifiers = { super_admin: 'phone_number', user: 'uuid' }`
   #
-  config.identifiers = { user: 'email' }
+  config.identifiers = {user: 'email'}
 
   # Configure the Devise trackable strategy integration.
   #
