@@ -1,16 +1,28 @@
 Rails.application.routes.draw do
-  
   apipie
+
+  resources :passive_workouts
+
+  resources :passive_activities
+
+  resources :passive_sleeps
+
+  resources :effective_time_intervals
+
+  resources :passive_data_points
+
+  resources :pre_defined_cards
+
   resources :reviews
 
   resources :rewards
 
-  resources :reminders, except: [:new, :destroy]
+  resources :reminders
 
-  resources :scorecards, only: [:index, :edit, :update]
-  
+  resources :scorecards
+
   devise_for :users, :path => '', :path_names => {:sign_in => 'login', :sign_out => 'logout'}, :controllers => {:registrations => 'my_devise/registrations',
-    :omniauth_callbacks => "omniauth_callbacks", :sessions => 'users/sessions', :passwords => 'users/passwords'}
+                                                                                                                :omniauth_callbacks => "omniauth_callbacks", :sessions => 'my_devise/sessions', :passwords => 'users/passwords', :invitations => 'my_devise/invitations'}
   match '/users/:id/finish_signup' => 'users#finish_signup', via: [:get, :patch], :as => :finish_signup
 
   # Auto-created by Devise
@@ -26,7 +38,11 @@ Rails.application.routes.draw do
 
   get 'superusers/index'
 
-  get 'pros/index'
+  resources :pros, only: [:index]
+
+  resources :supporters, only: [:index] do
+    post '/invite' => 'supporters#invite', :on => :collection
+  end
 
   match 'moods/cancel' => 'moods#cancel', :via => :get
 
@@ -52,7 +68,31 @@ Rails.application.routes.draw do
     get "delete"
   end
 
+  get "/update_capture" => 'capture#update_capture', as: 'update_capture'
+
+  get "/app_constants" => 'application#app_constants', as: 'app_constants'
+
+  get "/feature_flags" => 'feature_flags#feature_flags', as: 'feature_flags'
+
+  resources :conversations, :path => 'cards_stack', only: [:index, :show, :destroy] do
+    member do
+      post :reply
+      post :restore
+      post :mark_as_read
+    end
+    collection do
+      delete :empty_trash
+    end
+  end
+
+  resources :messages, :path => 'cards', only: [:new, :create] do
+    get "random_draw", :on => :collection
+  end
+
   resources :users, :path => 'thrivers' do
+    post "migrate_from_thrivetracker", :on => :collection
+    post "migrate_from_thrivetracker_admin", :on => :collection
+    post "request_password_reset_from_thrivetracker", :on => :collection
     resources :moods
     resources :sleeps
     resources :self_cares
@@ -60,14 +100,25 @@ Rails.application.routes.draw do
   end
 
   resources :connections, :controller => 'friendships', :except => [:show, :edit] do
-    get "requests", :on => :collection
-    get "invites", :on => :collection
+    get "thrivers", :on => :collection
+    get "supporters", :on => :collection
+    get "patients", :on => :collection
+    get "providers", :on => :collection
   end
+
+  resources :devices
+
+  flipper_constraint = lambda { |request| request.remote_ip == '127.0.0.1' }
+  constraints flipper_constraint do
+    mount Flipper::UI.app($flipper) => '/flipper'
+  end
+
+  get '/start' => 'high_voltage/pages#show', id: 'pricing'
 
   require 'api_constraints'
 
   # Api definition
-  namespace :api, defaults: { format: :json } do#, constraints: { subdomain: 'api' }, path: '/'  do
+  namespace :api, defaults: {format: :json} do #, constraints: { subdomain: 'api' }, path: '/'  do
     scope module: :v1, constraints: ApiConstraints.new(version: 1, default: true) do
       devise_scope :user do
         post '/registrations' => 'registrations#create'
