@@ -1,5 +1,5 @@
 class AddClientsToUsers < ActiveRecord::Migration
   def change
-  	add_column :users, :clients, :integer, array: true, default: []
+    add_column :users, :clients, :integer, array: true, default: []
   end
 end
